requirements.txt
Flask==2.3.2
firebase-admin==6.1.0
python-dotenv==1.0.0
flask-cors==4.0.0
requests==2.31.0
pytest==7.3.1
